+++ src/core/plugins/MidaPlugin.ts
@@ -3,39 +3,39 @@ import { MidaPluginParameters } from "#plugins/MidaPluginParameters";
 import { GenericObject } from "#utilities/GenericObject";
 
 export class MidaPlugin {
-    private readonly _id: string;
-    private readonly _name: string;
-    private readonly _description: string;
-    private readonly _version: string;
-    private readonly _install?: (actions: MidaPluginActions, options?: GenericObject) => void;
+    readonly #id: string;
+    readonly #name: string;
+    readonly #version: string;
+    readonly #description: string;
+    readonly #install?: (actions: MidaPluginActions, options?: GenericObject) => void;
 
-    public constructor ({ id, name, description = "", version, install, }: MidaPluginParameters) {
-        this._id = id;
-        this._name = name;
-        this._description = description;
-        this._version = version;
-        this._install = install;
+    public constructor ({ id, name, version, description = "", install, }: MidaPluginParameters) {
+        this.#id = id;
+        this.#name = name;
+        this.#version = version;
+        this.#description = description;
+        this.#install = install;
     }
 
     public get id (): string {
-        return this._id;
+        return this.#id;
     }
 
     public get name (): string {
-        return this._name;
-    }
-
-    public get description (): string {
-        return this._description;
+        return this.#name;
     }
 
     public get version (): string {
-        return this._version;
+        return this.#version;
+    }
+
+    public get description (): string {
+        return this.#description;
     }
 
     public install (actions: MidaPluginActions, options?: GenericObject): void {
-        if (this._install) {
-            this._install(actions, options);
+        if (this.#install) {
+            this.#install(actions, options);
         }
     }
 }