+++ src/core/symbols/MidaSymbolType.ts
@@ -1,8 +1,5 @@
 /** The type of a symbol. */
 export enum MidaSymbolType {
     FOREX = "forex",
-    CRYPTO = "crypto",
-    INDEX = "index",
     STOCK = "stock",
-    COMMODITY = "commodity",
 }